# perf: basic optimizations