--- a/types/tuple/mod.ts
+++ b/types/tuple/mod.ts
@@ -21,12 +21,13 @@ export class Tuple<
   }
 
   read(dataView: DataView, byteOffset = 0): V {
-    const tuple = [];
+    const len = this.typeOffsets.length;
 
-    for (const [typeOffset, type] of this.typeOffsets) {
-      tuple.push(type.read(dataView, byteOffset + typeOffset));
+    const tuple = new Array(len);
+    for (let i = 0; i < len; i++) {
+      const [typeOffset, type] = this.typeOffsets[i];
+      tuple[i] = type.read(dataView, byteOffset + typeOffset);
     }
-
     return tuple as V;
   }