Most of the benchmarks haven't decreased by much. But p995 and max have been reduces by a lot. Which should also help with scenario's where these functions are not jit compiled (--jitless or in cold paths)
### Struct Read
```
Original  119.36 ns/iter  (113.08 ns … 477.2 ns) 119.72 ns 139.11 ns 141.89 ns
Opt        96.91 ns/iter  (92.87 ns … 111.48 ns)  98.65 ns  102.8 ns 107.44 ns
```

### Struct Write
```
Original   99.87 ns/iter  (95.88 ns … 159.56 ns) 101.34 ns 113.85 ns  147.5 ns
Opt        68.71 ns/iter     (64 ns … 138.86 ns)  70.36 ns  81.27 ns 103.76 ns
```

### String Read
```
Original     1.28 µs/iter   (883.98 ns … 2.19 µs)   1.42 µs   2.19 µs   2.19 µs
Opt        178.88 ns/iter (174.17 ns … 315.51 ns) 180.14 ns 192.44 ns 247.79 ns
```

### Tuple Read
```
Original    35.36 ns/iter   (30.7 ns … 105.51 ns)  36.08 ns  67.82 ns  80.11 ns
Opt         20.99 ns/iter   (19.83 ns … 52.47 ns)  20.63 ns  26.84 ns  29.62 ns
```

### Tuple Write
```
Original    27.94 ns/iter   (26.97 ns … 65.92 ns)  27.92 ns   31.8 ns   43.1 ns
Opt         25.88 ns/iter   (24.34 ns … 37.55 ns)  25.95 ns  26.28 ns  26.65 ns
```