@MierenManz 2023-01-10
Further possible optimizations are to remove any array destructing. Once jitted it is as fast. But on sparkplug (baseline compiler) or intepreter it is slower

Small example here (run with `deno run --v8-flags=--jitless`
```ts
const arr = new Array(3).fill(null);

const start = performance.now();

for (let i = 0; i < 1_000_000; i++) {
  const [ a, b, c ] = arr;
}

const mid = performance.now();

for (let i = 0; i < 1_000_000; i++) {
  const a = arr[0],
  b = arr[1],
  c = arr[2];
}

const end = performance.now();

console.log({
  destructing: mid - start,
  indexing: end - mid,
})
```